libc: enable socket test
JIRA: CI-404
