+++ libc/test.yaml
@@ -25,7 +25,7 @@ test:
       execute: test-libc-pthread
       targets:
         include: [host-generic-pc]
-      
+
     - name: misc
       execute: test-libc-misc
       targets:
@@ -62,3 +62,15 @@ test:
       execute: test-libc-time
       targets:
         include: [host-generic-pc]
+
+    - name: unix-socket
+      execute: test-libc-unix-socket
+      targets:
+        # Excluded due to long execution time
+        exclude: [armv7m4-stm32l4x6-nucleo]
+        include: [host-generic-pc]
+
+    - name: inet-socket
+      execute: test-libc-inet-socket
+      targets:
+        value: [ia32-generic-qemu, armv7a7-imx6ull-evk, host-generic-pc]
