--- socket/Makefile
@@ -1,14 +1,0 @@
-#
-# Makefile for socket tests
-#
-# Copyright 2021 Phoenix Systems
-#
-# %LICENSE%
-#
-
-
-NAME := test-socket
-LOCAL_SRCS := test_socket.c
-DEP_LIBS := unity
-
-include $(binary.mk)
